Move chapter 6 exercises and solutions into standard layout.
