+++ src/test/kotlin/chapter6/exercises/ex1/listing.kt
@@ -1,25 +1,22 @@
-package chapter6.solutions
+package chapter6.exercises.ex1
 
 import chapter6.RNG
+import chapter6.unusedRng
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
 
-//tag::init[]
-fun nonNegativeInt(rng: RNG): Pair<Int, RNG> {
-    val (i1, rng2) = rng.nextInt()
-    return Pair(if (i1 < 0) -(i1 + 1) else i1, rng2)
-}
-//end::init[]
+/**
+ * TODO: Re-enable tests by removing `!` prefix!
+ */
+class Exercise1 : WordSpec({
 
-class Solution_6_1 : WordSpec({
+    //tag::init[]
+    fun nonNegativeInt(rng: RNG): Pair<Int, RNG> = TODO()
+    //end::init[]
 
     "nonNegativeInt" should {
 
-        val unusedRng = object : RNG {
-            override fun nextInt(): Pair<Int, RNG> = TODO()
-        }
-
-        "return 0 if nextInt() yields 0" {
+        "!return 0 if nextInt() yields 0" {
 
             val rng0 = object : RNG {
                 override fun nextInt(): Pair<Int, RNG> =
@@ -29,7 +26,7 @@ class Solution_6_1 : WordSpec({
             nonNegativeInt(rng0) shouldBe Pair(0, unusedRng)
         }
 
-        "return Int.MAX_VALUE when nextInt() yields Int.MAX_VALUE" {
+        "!return Int.MAX_VALUE when nextInt() yields Int.MAX_VALUE" {
 
             val rngMax = object : RNG {
                 override fun nextInt(): Pair<Int, RNG> =
@@ -42,7 +39,7 @@ class Solution_6_1 : WordSpec({
             )
         }
 
-        "return Int.MAX_VALUE when nextInt() yields Int.MIN_VALUE" {
+        "!return Int.MAX_VALUE when nextInt() yields Int.MIN_VALUE" {
 
             val rngMin = object : RNG {
                 override fun nextInt(): Pair<Int, RNG> =
@@ -55,7 +52,7 @@ class Solution_6_1 : WordSpec({
             )
         }
 
-        "return 0 when nextInt() yields -1" {
+        "!return 0 when nextInt() yields -1" {
 
             val rngNeg = object : RNG {
                 override fun nextInt(): Pair<Int, RNG> =
