+++ src/test/kotlin/chapter6/exercises/ex2/listing.kt
@@ -1,25 +1,22 @@
-package chapter6.solutions
+package chapter6.exercises.ex2
 
 import chapter6.RNG
+import chapter6.unusedRng
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
 
-//tag::init[]
-fun double(rng: RNG): Pair<Double, RNG> {
-    val (i, rng2) = nonNegativeInt(rng)
-    return Pair(i / (Int.MAX_VALUE.toDouble() + 1), rng2)
-}
-//end::init[]
+/**
+ * TODO: Re-enable tests by removing `!` prefix!
+ */
+class Exercise2 : WordSpec({
 
-class Solution_6_2 : WordSpec({
+    //tag::init[]
+    fun double(rng: RNG): Pair<Double, RNG> = TODO()
+    //end::init[]
 
     "double" should {
 
-        val unusedRng = object : RNG {
-            override fun nextInt(): Pair<Int, RNG> = TODO()
-        }
-
-        "generate a max value approaching 1 based on Int.MAX_VALUE" {
+        "!generate a max value approaching 1 based on Int.MAX_VALUE" {
 
             val rngMax = object : RNG {
                 override fun nextInt(): Pair<Int, RNG> =
@@ -32,7 +29,7 @@ class Solution_6_2 : WordSpec({
             )
         }
 
-        "generate a min value of 0 based on 0" {
+        "!generate a min value of 0 based on 0" {
             val rngMin = object : RNG {
                 override fun nextInt(): Pair<Int, RNG> =
                     Pair(0, unusedRng)
