+++ src/test/kotlin/chapter6/exercises/ex9/listing.kt
@@ -1,20 +1,19 @@
-package chapter6.solutions
+package chapter6.exercises.ex9
 
 import chapter6.Rand
-import chapter6.map
 import chapter6.rng1
 import chapter6.unit
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
 
-class Solution_6_9 : WordSpec({
+class Exercise9 : WordSpec({
+
     //tag::init1[]
-    fun <A, B> mapF(ra: Rand<A>, f: (A) -> B): Rand<B> =
-        flatMap(ra) { a -> unit(f(a)) }
+    fun <A, B> mapF(ra: Rand<A>, f: (A) -> B): Rand<B> = TODO()
     //end::init1[]
 
     "mapF" should {
-        "map over a value using flatMap" {
+        "!map over a value using flatMap" {
             mapF(
                 unit(1),
                 { a -> a.toString() })(rng1).first shouldBe "1"
@@ -29,16 +28,11 @@ class Solution_6_9 : WordSpec({
         ra: Rand<A>,
         rb: Rand<B>,
         f: (A, B) -> C
-    ): Rand<C> =
-        flatMap(ra) { a ->
-            map(rb) { b ->
-                f(a, b)
-            }
-        }
+    ): Rand<C> = TODO()
     //end::init2[]
 
     "map2F" should {
-        "combine the results of two actions" {
+        "!combine the results of two actions" {
 
             val combined: Rand<String> =
                 map2F(
