+++ src/test/kotlin/chapter6/solutions/sol5/listing.kt
@@ -1,18 +1,20 @@
-package chapter6.exercises
+package chapter6.solutions.sol5
 
 import chapter6.RNG
 import chapter6.Rand
+import chapter6.map
+import chapter6.solutions.sol1.nonNegativeInt
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
 
-/**
- * TODO: Re-enable tests by removing `!` prefix!
- */
-class Exercise_6_5 : WordSpec({
+//tag::init[]
+val doubleR: Rand<Double> =
+    map(::nonNegativeInt) { i ->
+        i / (Int.MAX_VALUE.toDouble() + 1)
+    }
+//end::init[]
 
-    //tag::init[]
-    val doubleR: Rand<Double> = TODO()
-    //end::init[]
+class Solution5 : WordSpec({
 
     "doubleR" should {
 
@@ -20,7 +22,7 @@ class Exercise_6_5 : WordSpec({
             override fun nextInt(): Pair<Int, RNG> = TODO()
         }
 
-        """!generate a max value approaching 1 based on
+        """generate a max value approaching 1 based on
             Int.MAX_VALUE using Rand""" {
 
                 val rngMax = object : RNG {
@@ -34,8 +36,7 @@ class Exercise_6_5 : WordSpec({
                 )
             }
 
-        "!generate a min value of 0 based on 0 using Rand" {
-
+        "generate a min value of 0 based on 0 using Rand" {
             val rngMin = object : RNG {
                 override fun nextInt(): Pair<Int, RNG> =
                     Pair(0, unusedRng)
