+++ src/test/kotlin/chapter6/solutions/sol6/listing.kt
@@ -1,4 +1,4 @@
-package chapter6.solutions
+package chapter6.solutions.sol6
 
 import chapter6.RNG
 import chapter6.Rand
@@ -16,7 +16,7 @@ fun <A, B, C> map2(ra: Rand<A>, rb: Rand<B>, f: (A, B) -> C): Rand<C> =
     }
 //end::init[]
 
-class Solution_6_6 : WordSpec({
+class Solution6 : WordSpec({
     "map2" should {
 
         "combine the results of two actions" {
