+++ src/test/kotlin/chapter6/solutions/sol9/listing.kt
@@ -1,19 +1,21 @@
-package chapter6.exercises
+package chapter6.solutions.sol9
 
 import chapter6.Rand
+import chapter6.map
 import chapter6.rng1
+import chapter6.solutions.sol8.flatMap
 import chapter6.unit
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
 
-class Exercise_6_9 : WordSpec({
-
+class Solution9 : WordSpec({
     //tag::init1[]
-    fun <A, B> mapF(ra: Rand<A>, f: (A) -> B): Rand<B> = TODO()
+    fun <A, B> mapF(ra: Rand<A>, f: (A) -> B): Rand<B> =
+        flatMap(ra) { a -> unit(f(a)) }
     //end::init1[]
 
     "mapF" should {
-        "!map over a value using flatMap" {
+        "map over a value using flatMap" {
             mapF(
                 unit(1),
                 { a -> a.toString() })(rng1).first shouldBe "1"
@@ -28,11 +30,16 @@ class Exercise_6_9 : WordSpec({
         ra: Rand<A>,
         rb: Rand<B>,
         f: (A, B) -> C
-    ): Rand<C> = TODO()
+    ): Rand<C> =
+        flatMap(ra) { a ->
+            map(rb) { b ->
+                f(a, b)
+            }
+        }
     //end::init2[]
 
     "map2F" should {
-        "!combine the results of two actions" {
+        "combine the results of two actions" {
 
             val combined: Rand<String> =
                 map2F(
